CatalogItemService added
