+++ src/BlazorAdmin/Pages/Index.razor
@@ -1,4 +1,5 @@
 @page "/"
+@using BlazorAdmin.Services
 
 <h1>Manage Product Catalog</h1>
