+++ src/BlazorAdmin/Services/CatalogItemService.CatalogItem.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace BlazorAdmin.Services
+{
+    public class CatalogItem
+    {
+        public int Id { get; set; }
+        public int CatalogTypeId { get; set; }
+        public int CatalogBrandId { get; set; }
+
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public decimal Price { get; set; }
+        public string PictureUri { get; set; }
+    }
+}
